refactor: Rename ng_compiler to ng_compiler_module and move from source_gen/ to compiler/module/.
PiperOrigin-RevId: 243871394
