+++ angular/lib/src/compiler/module/ng_compiler_module.dart
