+++ angular/lib/src/compiler/stylesheet_compiler/processor.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
 
 import 'package:build/build.dart';
+import 'package:angular/src/compiler/module/ng_compiler_module.dart';
 import 'package:angular/src/compiler/source_module.dart';
-import 'package:angular/src/source_gen/common/ng_compiler.dart';
 import 'package:angular/src/source_gen/common/url_resolver.dart';
 import 'package:angular_compiler/cli.dart';
